Move schemas to paths corresponding to their namespace
